Project/brats-mri/requirements.txt
lpips
monai
monai-generative